% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional_uncertainty.R
\name{caglm}
\alias{caglm}
\title{Calibrated inference for generalized linear models}
\usage{
caglm(formulas, family, data, target, ...)
}
\arguments{
\item{formulas}{A list of formulas}

\item{family}{a description of the error distribution and link function to be used in the model}

\item{data}{A dataframe containing the variables in the model}
}
\description{
Calibrated inference for generalized linear models
}
\examples{
n <- 1000
X <- rnorm(n)
Z1 <- rnorm(n)
Z2 <- rnorm(n)
logit <- 1*X + X^2 + Z1 + Z2
Y = rbinom(n, size = 1, prob = exp(logit)/(1+exp(logit)))
df <- data.frame(cbind(X,Y,Z1,Z2))
data <- as.data.frame(cbind(Y,X))
formulas <- list(Y~X, Y ~ X + I(X^2), Y ~ X + Z1, Y ~ X + Z2 + I(X^2), Y ~ X + Z1 + Z2 + I(X^2))
caglm(formulas, family = "binomial", data=data, target="X")
summary(glm(Y ~ X + Z1 + Z2 + I(X^2), family = "binomial", data=df))
}
