% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional_uncertainty.R
\name{distributional_seed}
\alias{distributional_seed}
\title{Function for creating distributional seeds}
\usage{
distributional_seed(n, delta)
}
\arguments{
\item{n}{The sample size}

\item{delta}{The pertubation parameter}
}
\description{
Function for creating distributional seeds
}
\examples{
d_seed <- distributional_seed(1000,10)
qqnorm(drnorm(d_seed))
# Compare to standard normal
qqnorm(rnorm(1000))
}
