% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional_uncertainty.R
\name{drbeta}
\alias{drbeta}
\title{Distributional random number generation (equivalent to rbeta)}
\usage{
drbeta(d_seed, shape1, shape2)
}
\arguments{
\item{d_seed}{A distributional seed as returned by distributional_seed()}

\item{shape1/shape2}{non-negative parameters of the Beta distribution}
}
\description{
Distributional random number generation (equivalent to rbeta)
}
\examples{
d_seed <- distributional_seed(1000,10)
drbeta(d_seed)
}
