% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional_uncertainty.R
\name{drbinom}
\alias{drbinom}
\title{Distributional random number generation (equivalent to rbinom)}
\usage{
drbinom(d_seed, size, prob)
}
\arguments{
\item{d_seed}{A distributional seed as returned by distributional_seed()}

\item{size}{number of trials}

\item{prob}{probability of success on each trial}
}
\description{
Distributional random number generation (equivalent to rbinom)
}
\examples{
d_seed <- distributional_seed(1000,10)
drbinom(d_seed)
}
