% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional_uncertainty.R
\name{drgamma}
\alias{drgamma}
\title{Distributional random number generation (equivalent to rgamma)}
\usage{
drgamma(d_seed, shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{d_seed}{A distributional seed as returned by distributional_seed()}

\item{shape}{shape parameter}

\item{rate}{an alternative way to specify the scale}

\item{scale}{scale parameter}
}
\description{
Distributional random number generation (equivalent to rgamma)
}
\examples{
d_seed <- distributional_seed(1000,10)
drgamma(d_seed)
}
