% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributional_uncertainty.R
\name{drunif}
\alias{drunif}
\title{Distributional random number generation (equivalent to runif)}
\usage{
drunif(d_seed, min = 0, max = 1)
}
\arguments{
\item{d_seed}{A distributional seed as returned by distributional_seed()}

\item{min}{lower limit of the unperturbed uniform distribution}

\item{max}{upper limit of the unperturbed uniform distribution}
}
\description{
Distributional random number generation (equivalent to runif)
}
\examples{
d_seed <- distributional_seed(1000,10)
drunif(d_seed)
}
